% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidDevTools-package}
\alias{AcidDevTools}
\alias{AcidDevTools-package}
\title{AcidDevTools}
\description{
Toolkit for R package development
}
\details{
Personal package with additional devtools utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/aciddevtools/}
  \item \url{https://github.com/acidgenomics/r-aciddevtools/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-aciddevtools/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
