% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{cd}
\alias{clear}
\alias{d}
\alias{la}
\title{Aliases}
\usage{
cd(...)

clear()

d(...)

la(...)
}
\arguments{
\item{...}{Passthrough arguments.}
}
\value{
Varies, depending on the function.
}
\description{
Mnemonic functions and shortcut aliases, to reduce common keystrokes.
}
\section{Functions}{
\itemize{
\item \code{cd()}: Alias for \code{setwd()}.

\item \code{clear()}: Alias for \code{cat("\\f")}.

\item \code{d()}: Alias for \code{document()}.

\item \code{la()}: Alias for \code{load_all()}.

}}
\note{
Updated 2021-08-23.
}
\examples{
## > d()
}
