% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb8.R
\name{bb8}
\alias{bb8}
\title{bb8}
\usage{
bb8()
}
\value{
Invisible character vector of packages attached specifically by
this function call.
}
\description{
Trusty sidekick for R package development.
}
\examples{
\dontrun{
bb8()
}
}
