% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build source and binary packages}
\usage{
build(package = getwd())
}
\arguments{
\item{package}{\code{character}.
Directory path to R package.}
}
\value{
\code{character}.
Named character vector containing \code{binary} and \code{source} file paths.
}
\description{
Build source and binary packages
}
\details{
Files get built into temporary directory, defined in \code{tempdir()}.
}
\note{
Updated 2022-05-23.
}
\examples{
## > package <- file.path("~", "monorepo", "r-packages", "r-goalie")
## > build(package = package)
}
\seealso{
\verb{R CMD build --help}.
}
