% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheTestFiles.R
\name{cacheTestFiles}
\alias{cacheTestFiles}
\title{Cache test files}
\usage{
cacheTestFiles(pkg, files)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package name.}

\item{files}{\code{character}.
File basenames relative to \code{remoteDir}.}
}
\value{
Invisible \code{character}.
File paths.
}
\description{
Cache test files
}
\note{
Updated 2023-09-28.
}
