% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catVec.R
\name{catVec}
\alias{catVec}
\title{Print (cat) a vector to console}
\usage{
catVec(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Console output.
}
\description{
Print (cat) a vector to console
}
\note{
Updated 2023-09-07.
}
\examples{
## Unnamed vector.
x <- c("aaa", "bbb", "ccc")
catVec(x)

## Named vector.
x <- c("aaa" = "AAA", "bbb" = "BBB", "ccc" = "CCC")
catVec(x)
}
