% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change directory}
\usage{
cd(dir)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory path.}
}
\description{
Alias for \code{\link[=setwd]{setwd()}} function.
}
\note{
Updated 2019-10-19.
}
\examples{
## > cd("~")
}
