% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearWarnings.R
\name{clearWarnings}
\alias{clearWarnings}
\title{Clear (reset) all warnings}
\usage{
clearWarnings()
}
\value{
No value.
}
\description{
Reassigns \code{last.warning} value in base environment.
}
\note{
Updated 2019-10-19.
}
\examples{
clearWarnings()
}
