% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentBiocVersion.R
\name{currentBiocVersion}
\alias{currentBiocVersion}
\title{Current Bioconductor release version}
\usage{
currentBiocVersion()
}
\value{
\code{numeric_version}.
}
\description{
Current Bioconductor release version
}
\note{
Updated 2020-11-11.
}
\examples{
currentBiocVersion()
}
\seealso{
\itemize{
\item https://bioconductor.org/bioc-version
\item https://bioconductor.org/config.yaml
}
}
