% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev}
\alias{dev}
\title{Load Developer Environment}
\usage{
dev()
}
\value{
Invisible character vector of packages attached specifically by
this function call.
}
\description{
Load Developer Environment
}
\examples{
\dontrun{
dev()
}
}
