% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\docType{package}
\name{devtoolsMJS-package}
\alias{devtoolsMJS}
\alias{devtoolsMJS-package}
\title{devtoolsMJS}
\description{
Personal package with additional devtools utilities.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
