% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{drat}
\alias{drat}
\title{Build packages and commit to drat repo}
\usage{
drat(
  package = getwd(),
  repo = file.path("~", "monorepo", "r-acidgenomics-com"),
  check = TRUE,
  tag = TRUE,
  pkgdown = TRUE,
  deploy = TRUE
)
}
\arguments{
\item{package}{\code{character}.
Directory paths to R package source code.
Vectorized, supporting the building of multiple packages in a single call.}

\item{repo}{\code{character(1)}.
Directory path to local R drat repository (e.g. \code{"r-acidgenomics-com"}).}

\item{check}{\code{logical(1)}.
Perform package checks prior to building package.}

\item{tag}{\code{logical(1)}.
Tag release on GitHub.}

\item{pkgdown}{\code{logical(1)}.
Build pkgdown website, if supported.}

\item{deploy}{\code{logical(1)}.
Deploy (push) the drat repo to AWS S3.}
}
\value{
\code{logical(1)}.
Boolean, indicating if drat build was successful.
}
\description{
Build packages and commit to drat repo
}
\note{
Updated 2022-10-20.
}
\section{Building from the command line}{


\if{html}{\out{<div class="sourceCode sh">}}\preformatted{R CMD build ~/monorepo/r-packages/r-syntactic
R CMD INSTALL --build syntactic_0.5.2.tar.gz
R CMD check --as-cran syntactic_0.5.2.tar.gz
# How to install from source:
R CMD INSTALL syntactic_0.5.2.tar.gz
# Or how to install binary (on macOS).
R CMD INSTALL syntactic_0.5.2.tgz
}\if{html}{\out{</div>}}
}

\section{Building vignettes}{


If you run into problems with a vignette index not being detected, make sure
the \code{.Rbuildignore} file does not contain \verb{^vignettes$}. Ignoring \verb{^Meta$}
and \verb{^doc$} is acceptable though.

Note that \code{devtools::build} currently doesn't handle building vignettes
for binary packages correctly.
}

\examples{
## > drat(
## >     package = file.path(
## >         "~",
## >         "monorepo",
## >         "r-packages",
## >         c(
## >             "r-acidbase",
## >             "r-basejump",
## >             "r-koopa"
## >         )
## >     )
## > )
}
\seealso{
\itemize{
\item https://seandavi.github.io/post/build-linux-r-binary-packages/
\item https://rstudio.github.io/r-manuals/r-exts/Creating-R-packages.html
}
}
