% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndReplace.R
\name{findAndReplace}
\alias{findAndReplace}
\title{Find and replace in files across a directory}
\usage{
findAndReplace(
  pattern,
  replacement,
  filePattern = "\\\\.(r|R)$",
  dir = getwd(),
  recursive = FALSE
)
}
\arguments{
\item{pattern}{\code{character(1)}.
Pattern string, supporting regular expressions.}

\item{replacement}{\code{character(1)}.
Replacement string.}

\item{filePattern}{\code{character(1)}.
File pattern matching string.
Defaults to matching against R files.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{recursive}{\code{logical(1)}.
Search recursively?}
}
\value{
Invisibly return file paths.
}
\description{
Find and replace in files across a directory
}
\note{
Updated 2023-05-22.
}
\examples{
unlink("testdata", recursive = TRUE)
dir.create(file.path("testdata", "subdir"), recursive = TRUE)
writeLines(
    text = "print(\"foo\")",
    con = file.path("testdata", "aaa.R"),
)
writeLines(
    text = "print(\"foo\")",
    con = file.path("testdata", "subdir", "bbb.R"),
)
out <- findAndReplace(
    pattern = "foo",
    replacement = "bar",
    dir = "testdata",
    recursive = TRUE
)
print(out)
print(readLines(out[[1L]]))
unlink("testdata", recursive = TRUE)
}
