% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{finder}
\alias{finder}
\title{Open Finder to the current directory.}
\usage{
finder(path = ".")
}
\arguments{
\item{path}{File path.}
}
\description{
Open Finder to the current directory.
}
\note{
macOS only.
}
