% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentGitHubVersion.R
\name{getCurrentGitHubVersion}
\alias{getCurrentGitHubVersion}
\title{Get current release version of package from GitHub}
\usage{
getCurrentGitHubVersion(repo)
}
\arguments{
\item{repo}{\code{character}.
GitHub repository ("user/package").
Vectorized, supporting multiple requests in a single call.}
}
\value{
\code{package_version}.
Package \code{repo} is defined in \code{names}.
}
\description{
Get current release version of package from GitHub
}
\note{
Updated 2021-08-23.
}
\examples{
repo <- paste("r-lib", c("rlang", "testthat"), sep = "/")
print(repo)
x <- getCurrentGitHubVersion(repo)
print(x)
}
