% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentVersion.R
\name{getCurrentVersion}
\alias{getCurrentVersion}
\title{Get current version of package from GitHub}
\usage{
getCurrentVersion(repo)
}
\arguments{
\item{repo}{\code{character(1)}.
GitHub repository ("<user>/<package>").}
}
\description{
Get current version of package from GitHub
}
\note{
Updated 2019-10-23.
}
\examples{
getCurrentVersion(repo = "acidgenomics/bb8").
}
