% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gitBranch}
\alias{gitBranch}
\alias{gitCurrentBranch}
\alias{gitDefaultBranch}
\title{Git branch strings}
\usage{
gitCurrentBranch(repo = getwd())

gitDefaultBranch(repo = getwd(), remote = "origin")
}
\arguments{
\item{repo}{\code{character(1)}.
Git repository directory path.}

\item{remote}{\code{character(1)}.
Remote name.}
}
\value{
\code{character(1)}.
Git branch name.
}
\description{
Git branch strings
}
\note{
Updated 2022-05-02.
}
\examples{
## > repo <- file.path("~", "git", "monorepo", "r-packages", "r-koopa")
## > gitCurrentBranch(repo)
## ## [1] "develop"
## > gitDefaultBranch(repo)
## ## [1] "main"
}
\seealso{
\itemize{
\item \code{koopa::git_default_branch} (shell).
\item \verb{git branch} (shell).
}
}
