% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectRow.R
\name{inspectRow}
\alias{inspectRow}
\title{Inspect a single (nested) row of a data frame}
\usage{
inspectRow(object, i)
}
\arguments{
\item{object}{Object.
Data structure that can be coerced to \code{data.frame}.}

\item{i}{\code{integer(1)} or \code{character(1)}.
Row position or name.}
}
\value{
Console output.
Invisibly returns dropped data structure.
}
\description{
Inspect a single (nested) row of a data frame
}
\note{
Updated 2023-12-13.
}
\examples{
object <- datasets::mtcars
inspectRow(object, i = 1L)
}
