% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAcidverse.R
\name{installAcidverse}
\alias{installAcidverse}
\title{Install Acid Genomics packages}
\usage{
installAcidverse(reinstall = FALSE)
}
\arguments{
\item{reinstall}{\code{logical(1)}.
Force reinstallation of any existing packages.}
}
\value{
Status from \code{install()}.
}
\description{
Install Acid Genomics packages
}
\note{
Updated 2022-04-15.
}
\examples{
## > installAcidverse()
}
