% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installBioconductor.R
\name{installBioconductor}
\alias{installBioconductor}
\title{Install Bioconductor}
\usage{
installBioconductor(version = Sys.getenv("BIOC_VERSION"))
}
\arguments{
\item{version}{\code{character(1)}.
Bioconductor release version.
If left unset, defaults to latest release.}
}
\value{
Status from \code{\link[BiocManager:install]{BiocManager::install()}}.
}
\description{
Install Bioconductor
}
\details{
\itemize{
\item Installs BiocManager package automatically, if necessary.
\item Calls \code{\link[BiocManager:install]{BiocManager::install()}} internally.
}
}
\note{
Updated 2020-11-03.
}
\examples{
## > installBioconductor()
}
\seealso{
\code{\link[BiocManager:version]{BiocManager::version()}}.
}
