% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDefaultPackages.R
\name{installDefaultPackages}
\alias{installDefaultPackages}
\title{Install recommended default R packages}
\usage{
installDefaultPackages(all = FALSE)
}
\arguments{
\item{all}{\code{logical(1)}.
Install additional extra packages.}
}
\description{
Install recommended default R packages
}
\note{
Updated 2020-11-03.
}
\section{Tricky packages}{

\itemize{
\item \strong{2020-08-24}: mgcv v1.8-32 won't build on macOS due to \verb{_kij} missing.
Can fix by installing binary package instead.
This issue is fixed in v1.8-33, on CRAN 2020-08-27.
\item \strong{2020-08-05}: rgdal v1.5-15 won't build on Debian.
Fixed with v1.5-16.
\item \strong{2020-08-11}: cpp11 v0.2.0 update is breaking tidyr.
See https://github.com/tidyverse/tidyr/issues/1024 for details.
}
}

\examples{
## > installDefaultPackages()
}
\seealso{
\itemize{
\item https://www.bioconductor.org/packages/release/BiocViews.html#___Software
}
}
