% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installedPackages.R
\name{installedPackages}
\alias{installedPackages}
\title{Find installed packages}
\usage{
installedPackages()
}
\description{
Includes "source" column, indicating whether package is from CRAN,
Bioconductor, or from a remote (i.e. GitHub, GitLab) install.
}
\note{
Updated 2019-10-17.
}
\examples{
x <- installedPackages()
table(x[["source"]], useNA = "ifany")
}
