% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGitHub.R
\name{installGitHub}
\alias{installGitHub}
\title{Install package from GitHub}
\usage{
installGitHub(repo, version = "latest", reinstall = FALSE)
}
\arguments{
\item{repo}{\code{character}.
Repository address(es) in the format \code{username/repo}.}
}
\value{
Invisible \code{character}.
Repository address defined in \code{repo} argument.
}
\description{
Install package from GitHub
}
\details{
This variant doesn't require \code{GITHUB_PAT}.
If you have a \code{GITHUB_PAT} defined, can use \code{\link[=install]{install()}} directly instead.
Intended for use inside container images, where a PAT may not be used.
}
\note{
Updated 2020-04-11.
}
\examples{
installGitHub(
    repo = c(
        "acidgenomics/acidbase",
        "acidgenomics/goalie"
    )
)
}
\seealso{
\itemize{
\item \code{remotes::install_github}.
\item \code{install.packages}.
}
}
