% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRecommendedPackages.R
\name{installRecommendedPackages}
\alias{installRecommendedPackages}
\title{Install recommended R packages}
\usage{
installRecommendedPackages()
}
\description{
Install recommended R packages
}
\note{
Updated 2022-06-09.
}
\examples{
## > installRecommendedPackages()
}
