% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCleanSystemLibrary.R
\name{isCleanSystemLibrary}
\alias{isCleanSystemLibrary}
\title{Is the R system library clean?}
\usage{
isCleanSystemLibrary()
}
\value{
\code{logical(1)}.
}
\description{
Determine whether a user has installed packages into the system library.
}
\note{
Updated 2020-08-11.
}
\examples{
isCleanSystemLibrary()
}
