% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInstalled.R
\name{isInstalled}
\alias{isInstalled}
\title{Is the desired package installed?}
\usage{
isInstalled(pkg)
}
\arguments{
\item{pkg}{\code{character}.
Package names.}
}
\value{
\code{logical(1)}.
}
\description{
Is the desired package installed?
}
\note{
Update 2019-10-23.
}
\examples{
isInstalled(pkg = c("bb8", "XXX"))
}
