% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\title{Kill current R session}
\usage{
kill()
}
\description{
Kill current R session
}
\note{
Updated 2019-10-22.
}
\examples{
## > kill()
}
\seealso{
\itemize{
\item RStudio: Session > Restart R (Ctrl + Shift + F10)
\item https://github.com/vsbuffalo/dotfiles/blob/master/.Rprofile
}

Useful kill commands in shell:
\itemize{
\item \code{kill -WINCH <PID>}.
\item \code{pkill rsession}.
\item \code{pgrep rsession}.
}
}
