% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrateRecommendedPackages.R
\name{migrateRecommendedPackages}
\alias{migrateRecommendedPackages}
\title{Migrate recommended packages}
\usage{
migrateRecommendedPackages()
}
\description{
Migrate recommended packages
}
\details{
Currently includes these packages: KernSmooth, MASS, Matrix, boot, class,
cluster, codetools, foreign, lattice, mgcv, nlme, nnet, rpart, spatial,
survival.
}
\note{
Updated 2022-09-06.
}
\seealso{
\itemize{
\item \code{.libPaths()}
\item \code{.Library}, \code{.Library.site}.
\item \code{.installed.packages()}, \code{install.packages()}.
}
}
