% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjs-package.R
\docType{package}
\name{mjs-package}
\alias{mjs}
\alias{mjs-package}
\title{MJS devtools}
\description{
Personal package with additional devtools utilities.
}
\author{
\strong{Maintainer}: Michael J. Steinbaugh \email{mike@steinbaugh.com}

}
