% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonCamelArgs.R
\name{nonCamelArgs}
\alias{nonCamelArgs}
\title{Return documentation with arguments that are not formatted in strict lower
camel case}
\usage{
nonCamelArgs(pkgName)
}
\arguments{
\item{pkgName}{\code{character(1)}.
Package name.}
}
\value{
\code{character}.
Offending Rd documentation files.
}
\description{
Return documentation with arguments that are not formatted in strict lower
camel case
}
\note{
Updated 2023-09-27.
}
