% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonCamelExports.R
\name{nonCamelExports}
\alias{nonCamelExports}
\title{Return exports that are not formatted in strict camel case}
\usage{
nonCamelExports(path = getwd())
}
\arguments{
\item{path}{\code{character(1)}.
File path (e.g. to package).}
}
\value{
\code{character}.
Names of exports that do not conform to strict camel case.
}
\description{
Return exports that are not formatted in strict camel case
}
\note{
Updated 2023-09-27.
}
\examples{
x <- nonCamelExports()
print(x)
}
