% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDependencies.R
\name{packageDependencies}
\alias{packageDependencies}
\title{Package dependencies}
\usage{
packageDependencies(pkg, recursive = TRUE)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package name.}

\item{recursive}{\code{logical(1)}.
Check dependencies recursively.}
}
\value{
\code{character}.
Names of dependency packages.
}
\description{
Package dependencies
}
\note{
Updated 2022-10-20.
}
\examples{
## > packageDependencies("stats")
}
