% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageProject.R
\name{packageProject}
\alias{packageProject}
\title{Package Project and Build Website}
\usage{
packageProject(install = FALSE)
}
\arguments{
\item{install}{Install package.}
}
\description{
Package Project and Build Website
}
