% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRd.R
\name{parseRd}
\alias{parseRd}
\alias{RdTags}
\title{Parse R documentation}
\usage{
parseRd(object, tag)

RdTags(object)
}
\arguments{
\item{object}{\code{Rd}.
R documentation, returned from \code{tools::Rd_db()}.}

\item{tag}{\code{character(1)}.
Desired metadata type.

These types are supported:
\itemize{
\item \code{title}.
\item \code{description}.
\item \code{usage}.
\item \code{arguments}.
\item \code{value}.
\item \code{references}.
\item \code{seealso}.
\item \code{examples}.
}}
}
\description{
Modified version of \code{tools:::.Rd_get_metadata()} that keeps whitespace and
returns \code{character} instead of \code{matrix}.
}
\section{Functions}{
\itemize{
\item \code{RdTags()}: Modified version of the unexported \code{tools:::RdTags()} function.

}}
\note{
Updated 2022-10-20.
}
\examples{
db <- tools::Rd_db("base")
head(names(db))
rd <- db[["nrow.Rd"]]
class(rd)
summary(rd)
RdTags(rd)
examples <- parseRd(rd, tag = "examples")
print(examples)
}
\seealso{
\code{tools::Rd_db()}.
}
