% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownDeployToAws.R
\name{pkgdownDeployToAws}
\alias{pkgdownDeployToAws}
\title{Deploy a pkgdown website to AWS S3 bucket}
\usage{
pkgdownDeployToAws(
  package = getwd(),
  bucketDir = "s3://r.acidgenomics.com/packages/",
  clean = TRUE
)
}
\arguments{
\item{package}{\code{character}.
Directory path to R package.
Parameterized, supporting building of multiple package websites in
a single call.}

\item{bucketDir}{\code{character(1)}.
AWS S3 bucket path.}

\item{clean}{\code{logical(1)}.
Ensure pkgdown website build is clean, and also deleted after upload
to AWS S3.}
}
\value{
\code{logical}.
Whether individual pkgdown builds were successful, corresponding to
input defined in \code{package} argument.
}
\description{
Deploy a pkgdown website to AWS S3 bucket
}
\note{
Updated 2023-09-27.
}
\examples{
## > pkgdownDeployToAws(
## >     package = file.path("~", "monorepo", "r-packages", "r-koopa")
## > )
}
