% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printComment.R
\name{printComment}
\alias{printComment}
\title{Print as comment}
\usage{
printComment(..., prefix = c("#", "# >", "#'", "#' >", "##", "## >"))
}
\arguments{
\item{...}{Passthrough to \code{print()}.}

\item{prefix}{Comment prefix to use. RStudio, roxygen, and ESS prefixes
are supported.}
}
\value{
Console output.
}
\description{
Print as comment
}
