% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Build packages and publish to Acid Genomics repo}
\usage{
publish(package = getwd(), check = TRUE, tag = TRUE, pkgdown = TRUE)
}
\arguments{
\item{package}{\code{character}.
Directory paths to R package source code.
Vectorized, supporting the building of multiple packages in a single call.}

\item{check}{\code{logical(1)}.
Perform package checks prior to building package.}

\item{tag}{\code{logical(1)}.
Tag release on GitHub.}

\item{pkgdown}{\code{logical(1)}.
Build pkgdown website, if supported.}
}
\value{
\code{logical(1)}.
Boolean, indicating if package build was successful.
}
\description{
Build packages and publish to Acid Genomics repo
}
\details{
Delegates build, drat insert, S3 sync, and CloudFront invalidation to
\verb{koopa app r publish}. Optional package checks and pkgdown site deployment
are handled in R before handing off to koopa.
}
\note{
Updated 2026-06-20.
}
\examples{
## > publish(
## >     package = file.path(
## >         "~",
## >         "monorepo",
## >         "r-packages",
## >         c(
## >             "r-acidbase",
## >             "r-basejump",
## >             "r-koopa"
## >         )
## >     )
## > )
}
