% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{biocValid}
\alias{build}
\alias{build_vignettes}
\alias{document}
\alias{install}
\alias{load_all}
\alias{test}
\alias{use_data}
\alias{use_data_raw}
\alias{use_testthat}
\alias{lint_package}
\alias{\%>\%}
\alias{build_site}
\title{Objects exported from other packages}
\description{
Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages

Objects exported from other packages
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocInstaller}{\code{\link[BiocInstaller]{biocValid}}}

  \item{devtools}{\code{\link[devtools]{build}}, \code{\link[devtools]{build_vignettes}}, \code{\link[devtools]{document}}, \code{\link[devtools]{install}}, \code{\link[devtools]{load_all}}, \code{\link[devtools]{test}}, \code{\link[devtools]{use_data}}, \code{\link[devtools]{use_data_raw}}, \code{\link[devtools]{use_testthat}}}

  \item{lintr}{\code{\link[lintr]{lint_package}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{pkgdown}{\code{\link[pkgdown]{build_site}}}
}}

