% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderProject.R
\name{renderProject}
\alias{renderProject}
\title{Render All RMarkdown Files in Working Directory}
\usage{
renderProject(today = TRUE)
}
\arguments{
\item{today}{Render to a subdirectory with today's date.}
}
\value{
No value.
}
\description{
Render All RMarkdown Files in Working Directory
}
