% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-scriptPath.R
\name{scriptPath}
\alias{scriptPath}
\title{Script path}
\usage{
scriptPath()
}
\value{
\code{character(1)}.
Normalized path to current script.
}
\description{
Script path
}
\note{
Only currently works inside RStudio.

Updated 2020-04-12.
}
