% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Friendly human readable object size}
\usage{
size(x)
}
\value{
\code{character(1)}.
Object size.
}
\description{
Friendly human readable object size
}
\note{
Updated 2023-08-09.
}
\examples{
x <- c("aaa", "bbb")
print(size(x))
}
