% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{styler}
\alias{styler}
\alias{acid_style}
\title{Acid Genomics styler formatting theme}
\usage{
acid_style()
}
\value{
\code{list}.
List containing styler rule definitions.
}
\description{
Acid Genomics styler formatting theme
}
\note{
Updated 2022-03-14.
}
\seealso{
\itemize{
\item \code{styler::tidyverse_style}.
\item https://styler.r-lib.org/articles/customizing_styler.html
--https://styler.r-lib.org/articles/remove_rules.html
\item https://github.com/r-lib/lintr/issues/43
}
}
