% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\title{R documentation table}
\usage{
tabular(x)
}
\arguments{
\item{x}{\code{data.frame}.}
}
\value{
Console output.
}
\description{
R documentation table
}
\note{
Updated 2019-08-13.
}
\examples{
df <- data.frame(
    "aaa" = seq(from = 1L, to = 4L),
    "bbb" = seq(from = 2L, to = 5L),
    "ccc" = seq(from = 3L, to = 6L),
    row.names = c("AAA", "BBB", "CCC", "DDD")
)
tabular(df)
}
\seealso{
\itemize{
\item http://r-pkgs.had.co.nz/man.html
}
}
