% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Execute test_that tests in a package}
\usage{
test(path = ".", coverage = TRUE)
}
\arguments{
\item{path}{\code{character(1)}.
File path (e.g. to package).}

\item{coverage}{\code{logical(1)}.
Check that unit tests provide sufficient coverage.}
}
\value{
Invisible \code{TRUE} on success, otherwise error.
}
\description{
Execute test_that tests in a package
}
\note{
Updated 2021-07-06.
}
\examples{
## > test()
}
