% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall a package}
\usage{
uninstall(pkgs)
}
\arguments{
\item{pkgs}{\code{character}.
Package names.}
}
\value{
Invisible \code{character}.
Names of packages that were removed.
}
\description{
Wrapper for \code{utils::remove.packages()}.
}
\note{
Updated 2020-04-09.
}
\examples{
## > uninstall("bb8")
}
