% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update all installed packages}
\usage{
update()
}
\value{
Invisible \code{logical(1)}.
Whether installation passes Bioconductor validity checks.
See \code{\link[BiocManager:valid]{BiocManager::valid()}} for details.
}
\description{
Update all installed packages
}
\note{
Conflicts with \code{stats::update()}.

Updated 2019-10-19.
}
\examples{
## > update()
}
