% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDeps.R
\name{updateDeps}
\alias{updateDeps}
\title{Update outdated dependencies}
\usage{
updateDeps(
  pkg = getwd(),
  type = c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")
)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package path. Must contain a \code{DESCRIPTION} file.}

\item{type}{\code{character}.
Dependency type.
See \code{utils::install.packages()} for details.}
}
\value{
\code{BiocManager::install()} call if packages need an update.
}
\description{
Supports Bioconductor, CRAN, and GitHub packages via BiocManager.
}
\details{
This is used by \href{https://github.com/acidgenomics/r-cmd-check}{r-cmd-check} for continuous integration (CI) testing.
}
\note{
Updated 2020-10-20.
}
\seealso{
\itemize{
\item desc package.
\item \code{utils::packageDescription()}.
}
}
