% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bb8-package}
\alias{bb8}
\alias{bb8-package}
\title{bb8}
\description{
Trusty sidekick for R package development
}
\details{
Personal package with additional devtools utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bb8.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/bb8/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
