% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bb8-package}
\alias{bb8}
\alias{bb8-package}
\title{bb8}
\description{
Trusty sidekick for R package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steinbaugh/bb8}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

}
