% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb8.R
\name{bb8}
\alias{bb8}
\title{Load Developer Environment}
\usage{
bb8()
}
\value{
Invisible character vector of packages attached specifically by
this function call.
}
\description{
Load Developer Environment
}
\examples{
\dontrun{
bb8()
}
}
