% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheTestFiles.R
\name{cacheTestFiles}
\alias{cacheTestFiles}
\title{Cache test files from Acid Genomics S3 bucket}
\usage{
cacheTestFiles(pkg, files)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package name.}

\item{files}{\code{character}.
File names in S3 testdata bucket.}
}
\value{
Invisible \code{list}.
List containing \code{"cacheDir"}, \code{"files"}, and \code{"remoteDir"}.
}
\description{
Cache test files from Acid Genomics S3 bucket
}
\note{
Updated 2023-09-28.
}
