% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catVec.R
\name{catVec}
\alias{catVec}
\title{Concatenate and Print a Vector}
\usage{
catVec(x)
}
\value{
Console output.
}
\description{
Print vector to console quoted, with comma delims
}
\examples{
c("hello", "world") \%>\% catVec()
}
