% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check package}
\usage{
check(path = ".", cran = FALSE)
}
\arguments{
\item{path}{\code{character(1)}.
Directory path to package.}

\item{cran}{\code{logical(1)}.
Perform additional CRAN submission checks.}
}
\value{
\code{TRUE} on success, otherwise error.
}
\description{
Check package
}
\note{
Updated 2021-05-18.
}
\examples{
## > check()
}
