% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev}
\alias{dev}
\title{Attach developer packages}
\usage{
dev(quiet = TRUE)
}
\arguments{
\item{quiet}{\code{logical(1)}.
Load packages quietly.}
}
\value{
Invisible character vector of packages attached specifically by
this function call.
}
\description{
Attach developer packages
}
\note{
Updated 2021-08-23.
}
\section{Current packages}{

\itemize{
\item magrittr
\item testthat
\item goalie
\item basejump
}
}

\examples{
## Load the developer environment.
## > dev()
}
