% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev}
\alias{dev}
\title{Attach developer packages}
\usage{
dev()
}
\value{
Invisible character vector of packages attached specifically by
this function call.
}
\description{
Attach developer packages
}
\examples{
## Load the developer environment.
## > bb8()
}
