% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacciSequence.R
\name{fibonacciSequence}
\alias{fibonacciSequence}
\title{Fibonacci sequence}
\usage{
fibonacciSequence(len)
}
\arguments{
\item{len}{\code{integer(1)}.}
}
\value{
\code{integer}.
}
\description{
Fibonacci sequence
}
\examples{
fibonacciSequence(8L)
}
