% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndReplace.R
\name{findAndReplace}
\alias{findAndReplace}
\title{Find and replace in files across a directory}
\usage{
findAndReplace(
  pattern,
  replacement,
  filePattern = "\\\\.(r|R)$",
  dir = ".",
  recursive = FALSE
)
}
\arguments{
\item{pattern}{\code{character(1)}.
Pattern string, supporting regular expressions.}

\item{replacement}{\code{character(1)}.
Replacement string.}

\item{filePattern}{\code{character(1)}.
File pattern matching string.
Defaults to matching against R files.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{recursive}{\code{logical(1)}.
Search recursively?}
}
\value{
Invisibly return file paths.
}
\description{
Find and replace in files across a directory
}
\note{
Updated 2021-08-23.
}
