% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndReplace.R
\name{findAndReplace}
\alias{findAndReplace}
\title{Find and replace across a directory.}
\usage{
findAndReplace(pattern, replacement, dir = ".", recursive = FALSE)
}
\arguments{
\item{pattern}{\code{character(1)}.
Pattern string.}

\item{replacement}{\code{character(1)}.
Replacement string.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{recursive}{\code{logical(1)}.
Search recursively?}
}
\value{
Invisibly return file paths.
}
\description{
Find and replace across a directory.
}
\note{
Updated 2019-08-13.
}
