% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos-finder.R
\name{finder}
\alias{finder}
\title{Open Finder to the current directory}
\usage{
finder(path = getwd())
}
\arguments{
\item{path}{\code{character(1)}.
File path (e.g. to package).}
}
\value{
Opens window. No return.
}
\description{
Open Finder to the current directory
}
\note{
macOS only.

Updated 2021-08-22.
}
\examples{
## > finder(path = "~")
}
