% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finder.R
\name{finder}
\alias{finder}
\title{Open Finder to the current directory}
\usage{
finder(path = ".")
}
\arguments{
\item{path}{\code{character(1)}.
File path.}
}
\value{
Opens window. No return.
}
\description{
Open Finder to the current directory
}
\note{
macOS only.

Updated 2020-04-07.
}
