% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentGitHubVersion.R
\name{getCurrentGitHubVersion}
\alias{getCurrentGitHubVersion}
\title{Get current version of package from GitHub}
\usage{
getCurrentGitHubVersion(repo)
}
\arguments{
\item{repo}{\code{character(1)}.
GitHub repository ("user/package").}
}
\description{
Get current version of package from GitHub
}
\note{
Updated 2020-07-01.
}
\examples{
getCurrentGitHubVersion(repo = "acidgenomics/bb8")
}
