% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages from Bioconductor, CRAN, or a Git remote}
\usage{
install(pkgs, dependencies = NA, lib = .libPaths()[[1L]], reinstall = TRUE)
}
\arguments{
\item{pkgs}{\code{character}.
Package names to install.
By default, strings are passed to \code{BiocManager::install()}.

Special cases:
\itemize{
\item Package tarball files and remote URLs (i.e. from CRAN or Bioconductor)
are supported.
\item Strings matching  "USER/REPO" are treated as GitHub repositories,
and installed using \code{remotes::install_github()}.
\item Strings ending with ".git" are treated as Git repositories, and
installed using \code{remotes::install_git()}.
}}

\item{dependencies}{\code{logical(1)}, \code{character}, or \code{NA}.
\itemize{
\item \code{TRUE}/\code{FALSE} indicating whether to install uninstalled packages which
these packages depend on/link to/import/suggest.
\item Can pass a \code{character} vector, a subset of
\code{c("Depends", "Imports", " LinkingTo", "Suggests", "Enhances")}.
\item Can pass \code{NA}, the default for
\code{\link[utils:install.packages]{install.packages()}}, which means
\code{c("Depends", "Imports", "LinkingTo")}.
}}

\item{lib}{\code{character}.
Destination library directory path.
Defaults to the first element of \code{.libPaths()}.}

\item{reinstall}{\code{logical(1)}.
Force reinstallation of any existing packages.}
}
\value{
Invisible \code{list}.
Contains information on \code{pkgs} and \code{lib} defined.
}
\description{
Install packages from Bioconductor, CRAN, or a Git remote
}
\note{
Updated 2022-07-28.
}
\examples{
## > testlib <- file.path(tempdir(), "testlib")
## > unlink(testlib, recursive = TRUE)
## > out <- install(
## >     pkgs = "BiocGenerics",
## >     dependencies = FALSE,
## >     lib = file.path(tempdir(), "testlib")
## > )
## > print(out)
## > list.dirs(path = testlib, full.names = FALSE, recursive = FALSE)
## > unlink(testlib, recursive = TRUE)
}
