% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages from Bioconductor, CRAN, or a Git remote}
\usage{
install(
  pkgs,
  configureArgs = getOption("configure.args"),
  configureVars = getOption("configure.vars"),
  dependencies = NA,
  lib = .libPaths()[[1L]],
  type = getOption("pkgType"),
  reinstall = TRUE
)
}
\arguments{
\item{pkgs}{\code{character}.
Package names to install.
Package names containing a "/" are treated as GitHub repositories and
installed using \code{\link[remotes:install_github]{remotes::install_github()}}.}

\item{configureArgs, configureVars}{\code{character} or named \code{list}.
\emph{Used only for source installs.} If a character vector with no names is
supplied, the elements are concatenated into a single string (separated by
a space) and used as the value for the \code{--configure-args} or
\code{configure-vars} flag in the call to \verb{R CMD INSTALL}. If the character
vector has names, these are assumed to identify values for
\code{--configure-args} or \code{--configure-vars} for individual packages. This
allows one to specify settings for an entire collection of packages which
will be used if any of those packages are to be installed.

A named list can be used also to the same effect, and that allows
multi-element character strings for each package which are concatenated to
a single string to be used as the value for \code{--configure-args} and/or
\code{--configure-vars}.}

\item{dependencies}{\code{logical(1)}, \code{character}, or \code{NA}.
\itemize{
\item \code{TRUE}/\code{FALSE} indicating whether to install uninstalled packages which
these packages depend on/link to/import/suggest.
\item Can pass a \code{character} vector, a subset of
\code{c("Depends", "Imports", " LinkingTo", "Suggests", "Enhances")}.
\item Can pass \code{NA}, the default for
\code{\link[utils:install.packages]{install.packages()}}, which means
\code{c("Depends", "Imports", "LinkingTo")}.
}}

\item{lib}{\code{character}.
Destination library directory paths.
Defaults to the first element of \code{.libPaths()}.}

\item{type}{\code{character(1)}.
Type of package to download and install. \code{"source"} is recommended by
default, but \code{"binary"} can be used on macOS or Windows to install
pre-built binaries from CRAN or Bioconductor.

Note that installing from source requires the correct GCC and GNU Fortran
binaries to be installed, and Apple LLVM/Clang compilers should not be used
on macOS. Refer to \url{macOS development tools} for details.}

\item{reinstall}{\code{logical(1)}.
Force reinstallation of any existing packages.}
}
\value{
Invisible \code{character}.
Package names defined in the \code{pkgs} argument.
}
\description{
Install packages from Bioconductor, CRAN, or a Git remote
}
\note{
Updated 2020-08-11.
}
\examples{
## > install(pkgs = c("DESeq2", "edgeR"))
}
