% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installedPackages.R
\name{installedPackages}
\alias{installedPackages}
\title{Find installed packages}
\usage{
installedPackages()
}
\description{
Includes "source" column, indicating whether package is from CRAN,
Bioconductor, or from a remote (i.e. GitHub, GitLab) install.
}
\note{
Updated 2020-04-12.
}
\examples{
x <- installedPackages()
table(x[["source"]])
}
\seealso{
\itemize{
\item \code{sessioninfo::package_info()}.
\item \code{utils::installed.packages()}.
}
}
