% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRecommendedPackages.R
\name{installRecommendedPackages}
\alias{installRecommendedPackages}
\title{Install recommended R packages}
\usage{
installRecommendedPackages(
  extra = c(acidverse = TRUE, alignment = FALSE, annotation = FALSE, cancer = FALSE,
    chipseq = FALSE, devtools = TRUE, enrichment = FALSE, filesystem = TRUE, general =
    TRUE, graphics = TRUE, riboseq = FALSE, rnaseq = TRUE, shiny = TRUE, singlecell =
    FALSE, smallrna = FALSE, statistics = TRUE, tidyverse = TRUE, variation = FALSE)
)
}
\arguments{
\item{extra}{\code{logical} or \code{logical(1)}.
Named logical vector of which types of extra packages to install.
Alternatively, can install all extra packages by setting to \code{TRUE}, or
disable the installation of any extra packages by setting to \code{FALSE}.}
}
\description{
Install recommended R packages
}
\note{
Updated 2021-05-12.
}
\section{tidyverse packages}{


The \href{https://www.tidyverse.org/packages/}{tidyverse} is a data science meta-package containing: DBI, dplyr,
forcats, ggplot2, googledrive, googlesheets4, haven, httr, jsonlite,
lubridate, purrr, readr, readxl, rvest, stringr, tibble, tidyr, and xml2.
}

\examples{
## > installRecommendedPackages()
}
