% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\alias{killAll}
\title{Kill R session}
\usage{
kill(pid = Sys.getpid())

killAll()
}
\arguments{
\item{pid}{\code{integer(1)}.
Process ID.}
}
\description{
Kill R session
}
\note{
Updated 2019-10-23.
}
\examples{
## > kill()
## > killAll()
}
\seealso{
\itemize{
\item RStudio: Session > Restart R (Ctrl + Shift + F10)
\item https://github.com/vsbuffalo/dotfiles/blob/master/.Rprofile
}

Useful kill commands in shell:
\itemize{
\item \code{pgrep rsession}.
\item \code{pkill rsession}.
\item \code{kill -WINCH <PID>}.
}
}
