% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint_dir.R
\name{lint_dir}
\alias{lint_dir}
\title{Lint a directory}
\usage{
lint_dir(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{\code{character(1)}.
Directory path.}

\item{recursive}{\code{logical(1)}.
Search recursively.}
}
\value{
Lintr checks.
}
\description{
Lint a directory
}
\note{
Updated 2019-08-05.
}
