% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linux-memfree.R
\name{memfree}
\alias{memfree}
\title{Get free memory statistics.}
\usage{
memfree()
}
\description{
Get free memory statistics.
}
\note{
Currently this only works for Linux.

Updated 2019-10-19.
}
\seealso{
\itemize{
\item \code{help(topic = "Memory", package = "base")}.
\item \code{help(topic = "Memory-limits", package = "base")}.
\item \code{utils:::format.object_size()}.
\item \code{print:::print.bytes()}.
\item https://stackoverflow.com/a/6457769
\item https://stackoverflow.com/a/29787527
\item https://stat.ethz.ch/R-manual/R-devel/library/base/html/Memory-limits.html
\item http://adv-r.had.co.nz/memory.html
}
}
