% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjs-package.R
\docType{package}
\name{mjs-package}
\alias{mjs}
\alias{mjs-package}
\title{MJS devtools}
\description{
Personal package with additional devtools utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjsteinbaugh/devtools}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

}
