% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonCamelArgs.R
\name{nonCamelArgs}
\alias{nonCamelArgs}
\title{Return documentation with arguments that are not formatted in strict lower
camel case}
\usage{
nonCamelArgs(
  pkgName,
  allowlist = c("FUN", "MARGIN", "USE.NAMES", "X", ".xname")
)
}
\arguments{
\item{pkgName}{\code{character(1)}.
Package name.}

\item{allowlist}{\code{character}.
Character vector of common non-camel argument names to allow.}
}
\value{
\code{character}.
Offending Rd documentation files.
}
\description{
Return documentation with arguments that are not formatted in strict lower
camel case
}
\note{
Updated 2023-09-27.
}
