% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDependencies.R
\name{packageDependencies}
\alias{packageDependencies}
\title{Package dependencies}
\usage{
packageDependencies(pkg)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package name.}
}
\value{
\code{character}.
Names of dependency packages.
}
\description{
Package dependencies
}
\note{
Updated 2022-04-29.
}
\examples{
packageDependencies("stats")
}
