% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{dir}{\code{character(1)}.
Directory path.}

\item{path}{\code{character(1)}.
File path.}

\item{pkgs}{\code{character}.
Package names.}

\item{object}{Object.}

\item{reinstall}{\code{logical(1)}.
Force reinstallation of any existing packages.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
