% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos-pbcopy.R
\name{pbcopy}
\alias{pbcopy}
\title{Copy to clipboard}
\usage{
pbcopy(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Copy to clipboard
}
\note{
Only works on macOS.

Updated 2019-10-19.
}
\examples{
x <- "hello world"
pbcopy(x)
}
