% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printComment.R
\name{printComment}
\alias{printComment}
\alias{pc}
\title{Print as comment}
\usage{
printComment(
  ...,
  prefix = c("##", "## >", "#'", "#' >", "#", "# >"),
  width = 80L
)

pc(...)
}
\arguments{
\item{...}{Passthrough to \code{print()}.}

\item{prefix}{Comment prefix to use. RStudio, roxygen, and ESS prefixes
are supported.}

\item{width}{\code{integer(1L)}.
Desired output width.
Defaults to 80 characters.}
}
\value{
Console output.
}
\description{
Print as comment
}
\note{
Updated 2020-12-16.
}
