% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildBinary.R
\name{rebuildBinary}
\alias{rebuildBinary}
\title{Rebuild a binary tarball from a source tarball}
\usage{
rebuildBinary(
  packageName,
  repo = file.path("~", "monorepo", "r-acidgenomics-com")
)
}
\arguments{
\item{packageName}{\code{character}.
Package name(s).}

\item{repo}{\code{character(1)}.
Drat repo directory path.}
}
\value{
Invisible output.
}
\description{
Rebuild a binary tarball from a source tarball
}
\note{
Updated 2023-10-24.
}
\examples{
## > rebuildBinary(c("AcidSingleCell", "DESeqAnalysis"))
}
\seealso{
\itemize{
\item \code{build()}.
\item \code{publish()}.
}
}
