% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{biocValid}
\alias{reexports}
\alias{build}
\alias{reexports}
\alias{build_vignettes}
\alias{reexports}
\alias{document}
\alias{reexports}
\alias{install}
\alias{reexports}
\alias{load_all}
\alias{reexports}
\alias{test}
\alias{reexports}
\alias{use_data}
\alias{reexports}
\alias{use_data_raw}
\alias{reexports}
\alias{use_testthat}
\alias{reexports}
\alias{lint_package}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocInstaller}{\code{\link[BiocInstaller]{biocValid}}}

  \item{devtools}{\code{\link[devtools]{build}}, \code{\link[devtools]{build_vignettes}}, \code{\link[devtools]{document}}, \code{\link[devtools]{install}}, \code{\link[devtools]{load_all}}, \code{\link[devtools]{test}}, \code{\link[devtools]{use_data}}, \code{\link[devtools]{use_data_raw}}, \code{\link[devtools]{use_testthat}}}

  \item{lintr}{\code{\link[lintr]{lint_package}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

