% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{BiocCheck}
\alias{package_coverage}
\alias{percent_coverage}
\alias{report}
\alias{build_vignettes}
\alias{document}
\alias{load_all}
\alias{run_examples}
\alias{lint}
\alias{lint_dir}
\alias{lint_package}
\alias{build_articles}
\alias{build_home}
\alias{build_news}
\alias{build_reference}
\alias{build_reference_index}
\alias{build_site}
\alias{rcmdcheck}
\alias{style_dir}
\alias{style_file}
\alias{style_pkg}
\alias{test_file}
\alias{use_data}
\alias{doc}
\alias{exit}
\title{Reexported functions}
\description{
Reexported functions
}
\note{
Updated 2020-07-01.
}
