% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{BiocCheck}
\alias{valid}
\alias{available}
\alias{export}
\alias{import}
\alias{loadData}
\alias{saveData}
\alias{report}
\alias{build}
\alias{build_vignettes}
\alias{document}
\alias{load_all}
\alias{run_examples}
\alias{test}
\alias{lint_dir}
\alias{lint_package}
\alias{build_articles}
\alias{build_home}
\alias{build_news}
\alias{build_reference}
\alias{build_reference_index}
\alias{build_site}
\alias{rcmdcheck}
\alias{render}
\alias{camelCase}
\alias{makeNames}
\alias{snakeCase}
\alias{use_data}
\alias{check}
\alias{devinstall}
\alias{doc}
\title{Reexported functions}
\description{
Intended for use inside Rprofile internal env.
}
\note{
Updated 2019-10-20.
}
