% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{BiocCheck}
\alias{valid}
\alias{available}
\alias{package_coverage}
\alias{percent_coverage}
\alias{report}
\alias{build}
\alias{build_vignettes}
\alias{document}
\alias{load_all}
\alias{run_examples}
\alias{gh}
\alias{lint}
\alias{lint_dir}
\alias{lint_package}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{with_parameters_test_that}
\alias{export}
\alias{import}
\alias{loadData}
\alias{saveData}
\alias{object_size}
\alias{build_articles}
\alias{build_home}
\alias{build_news}
\alias{build_reference}
\alias{build_reference_index}
\alias{build_site}
\alias{rcmdcheck}
\alias{download_version}
\alias{install_bioc}
\alias{install_bitbucket}
\alias{install_deps}
\alias{install_git}
\alias{install_github}
\alias{install_gitlab}
\alias{install_url}
\alias{py_config}
\alias{py_module_available}
\alias{use_virtualenv}
\alias{virtualenv_list}
\alias{render}
\alias{camelCase}
\alias{kebabCase}
\alias{makeNames}
\alias{snakeCase}
\alias{test_file}
\alias{use_data}
\alias{check}
\alias{devinstall}
\alias{doc}
\title{Reexported functions}
\note{
Updated 2020-07-01.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}
}}

