% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{BiocCheck}
\alias{package_coverage}
\alias{percent_coverage}
\alias{report}
\alias{build}
\alias{build_vignettes}
\alias{document}
\alias{load_all}
\alias{run_examples}
\alias{gh}
\alias{lint}
\alias{lint_dir}
\alias{lint_package}
\alias{loadData}
\alias{saveData}
\alias{build_articles}
\alias{build_home}
\alias{build_news}
\alias{build_reference}
\alias{build_reference_index}
\alias{build_site}
\alias{rcmdcheck}
\alias{download_version}
\alias{install_bioc}
\alias{install_bitbucket}
\alias{install_deps}
\alias{install_git}
\alias{install_github}
\alias{install_gitlab}
\alias{install_url}
\alias{py_config}
\alias{py_module_available}
\alias{use_virtualenv}
\alias{virtualenv_list}
\alias{render}
\alias{style_dir}
\alias{style_file}
\alias{style_pkg}
\alias{camelCase}
\alias{kebabCase}
\alias{makeNames}
\alias{snakeCase}
\alias{test_file}
\alias{view}
\alias{use_data}
\alias{devinstall}
\alias{doc}
\alias{exit}
\title{Reexported functions}
\description{
Reexported functions
}
\note{
Updated 2020-07-01.
}
