% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Execute testthat tests in a package}
\usage{
test(path = getwd())
}
\arguments{
\item{path}{\code{character(1)}.
File path (e.g. to package).}
}
\value{
Invisible \code{TRUE} on success, otherwise error.
}
\description{
Execute testthat tests in a package
}
\details{
Note that usage of \code{test_local} is now recommended instead of \code{test_dir},
and appears to be required for testthat checks to run successfully in
parallel.
}
\note{
Updated 2023-08-15.
}
\examples{
## > test()
}
\seealso{
\itemize{
\item https://github.com/r-lib/testthat/issues/1216
}
}
