% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall a package}
\usage{
uninstall(pkgs, lib = .libPaths()[[1L]])
}
\arguments{
\item{pkgs}{\code{character}.
Package names.}

\item{lib}{\code{character}.
Destination library directory path.
Defaults to the first element of \code{.libPaths()}.}
}
\value{
Invisible \code{list}.
Metadata containing names of packages that were removed, and the package
library path.
}
\description{
Wrapper for \code{utils::remove.packages()}.
}
\note{
Updated 2022-10-20.
}
\examples{
## > uninstall("rlang")
}
