% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDeps.R
\name{updateDeps}
\alias{updateDeps}
\title{Update outdated dependencies}
\usage{
updateDeps(pkg = ".", type = c("Depends", "Imports"))
}
\arguments{
\item{pkg}{\code{character(1)}.
Package path. Must contain a \code{DESCRIPTION} file.}

\item{type}{\code{character}.
Package type.}
}
\value{
\code{BiocManager::install()} call if packages need an update.
}
\description{
Supports Bioconductor, CRAN, and GitHub packages via BiocManager.
}
\note{
Updated 2019-08-15.
}
