% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDeps.R
\name{updateDeps}
\alias{updateDeps}
\title{Update outdated dependencies}
\usage{
updateDeps(pkg = ".", verbose = FALSE)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package path. Must contain a \code{DESCRIPTION} file.}

\item{verbose}{\code{logical(1)}.}
}
\value{
\code{BiocManager::install()} call if packages need an update.
}
\description{
Supports Bioconductor, CRAN, and GitHub packages via BiocManager.
}
\note{
Updated 2019-08-13.
}
