% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackages.R
\name{updatePackages}
\alias{updatePackages}
\title{Update all installed packages}
\usage{
updatePackages(lib = .libPaths()[[1L]])
}
\arguments{
\item{lib}{\code{character}.
Destination library directory path.
Defaults to the first element of \code{.libPaths()}.}
}
\value{
Invisible \code{TRUE} or console output.
Whether installation passes Bioconductor validity checks.
See \code{BiocManager::valid()} for details.
}
\description{
Update all installed packages
}
\note{
Updated 2022-05-31.
}
\examples{
## > updatePackages()
}
