% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackages.R
\name{updatePackages}
\alias{updatePackages}
\title{Update all installed packages}
\usage{
updatePackages()
}
\value{
Invisible \code{TRUE} or console output.
Whether installation passes Bioconductor validity checks.
See \code{\link[BiocManager:valid]{BiocManager::valid()}} for details.
}
\description{
Update all installed packages
}
\note{
Updated 2020-11-25.
}
\examples{
## > updatePackages()
}
