% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{Validate installed package versions against correct versions}
\usage{
valid(...)
}
\arguments{
\item{...}{Passthrough arguments to \code{BiocManager::valid()}.}
}
\value{
\code{logical(1)}.
}
\description{
Validate installed package versions against correct versions
}
\note{
Updated 2023-01-09.
}
\examples{
## > valid()
}
