% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/View-methods.R
\docType{methods}
\name{View}
\alias{View}
\alias{View,ANY-method}
\alias{View,DataFrame-method}
\title{Invoke a Data Viewer}
\usage{
View(x, ...)

\S4method{View}{ANY}(x, title)

\S4method{View}{DataFrame}(x, title)
}
\arguments{
\item{x}{an \R object which can be coerced to a data frame with
    non-zero numbers of rows and columns.}

\item{title}{title for viewer window.  Defaults to name of \code{x}
    prefixed by \code{Data:}.}
}
\description{
Invoke a Data Viewer
}
