% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-view2.R
\name{view2}
\alias{view2}
\title{View object in RStudio viewer}
\usage{
view2(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Invisible \code{NULL}.
This function puts up a window and returns immediately.
In RStudio, this will open a tabbed window.
}
\description{
This variant performs internal coercion to data.frame, so we can easily
view S4 DataFrame and GenomicRanges objects.
}
\note{
Updated 2021-10-14.
}
\examples{
## > view2(mtcars)
}
