% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewDF.R
\name{viewDF}
\alias{viewDF}
\title{View Data Frame}
\usage{
viewDF(object)
}
\value{
No return.
}
\description{
Utility function for quickly viewing an object as a \link{data.frame}.
}
\details{
This is particularly useful for opening S4 class \link{DataFrame} objects in
RStudio using the nicer \link{data.frame} tabbed graphical interface.
}
